pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-langchain-chatbot"
version = "0.1.0"
description = "FastAPI LangChain Chatbot with multi-provider LLM support, conversation management, MCP server integration, and automatic chart generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.0.350",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "sqlalchemy>=2.0.23",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.2",
    "websockets>=12.0",
    "openai>=1.6.0",
    "anthropic>=0.8.0",
    "langchain-mcp-adapters>=0.0.6",
    "langgraph>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.25.2",
    "coverage>=7.3.0",
]
